FileBackend class, factor out B-generic code
Much of the interesting code is independent of the Backend.  Separate
this code so it's not made more difficult to understand by CPP.  Some of
the "generic" code actually depends on fairly trivial details of the
Backend - the steps to get from Result to an actual file on disk.
Abstract that with FileBackend.